refactor: add player spawn
